requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
mysql-connector-python==9.2.0
pillow==11.1.0
sqlparse==0.5.3
